requirements.txt
streamlit
python-dotenv

langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
langchain-groq

sentence-transformers
faiss-cpu

youtube-transcript-api
pytube
pypdf

huggingface-hub
